requirements-dev.txt
black==24.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
PyJWT==2.8.0
SQLAlchemy==2.0.27
tomli==2.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
